Initial Commit after starting the Repo with the tests already passing.
Note: I do not like how we have to keep casting things from runes to strings.
Further I don't like the complexity of using the range but not using the
index because we have to calculate the index we care about as to avoid
including non-alphanumeric characters.  I'm also unhappy I had to make a
regex to test for alphanumerics because it feels like I'm using a bazooka
against an ant.